HU07 Add new test for saveReservation method and the tests were organized
